move pid file to it's own file
