+++ cmd/up/up_test.go
@@ -15,13 +15,8 @@ package up
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"strconv"
 	"testing"
 
-	"github.com/okteto/okteto/pkg/config"
 	"github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/model"
 )
@@ -103,33 +98,6 @@ func Test_printDisplayContext(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			printDisplayContext(tt.dev)
 		})
-	}
-
-}
-
-func TestCreatePIDFile(t *testing.T) {
-	deploymentName := "deployment"
-	namespace := "namespace"
-	if err := createPIDFile(namespace, deploymentName); err != nil {
-		t.Fatal("unable to create pid file")
-	}
-
-	filePath := filepath.Join(config.GetDeploymentHome(namespace, deploymentName), "okteto.pid")
-	if _, err := os.Stat(filePath); os.IsNotExist(err) {
-		t.Fatal("didn't create pid file")
-	}
-
-	filePID, err := ioutil.ReadFile(filePath)
-	if err != nil {
-		t.Fatal("pid file is corrupted")
-	}
-	if string(filePID) != strconv.Itoa(os.Getpid()) {
-		t.Fatal("pid file content is invalid")
-	}
-
-	cleanPIDFile(namespace, deploymentName)
-	if _, err := os.Create(filePath); os.IsExist(err) {
-		t.Fatal("didn't delete pid file")
 	}
 
 }
